Revert "UPSTREAM: sched/uclamp: Reject negative values in cpu_uclamp_write()"
This reverts commit 7230150060d70c87eaba39e8d04cd72f2fb4b29d.

+++ kernel/sched/core.c
@@ -8044,7 +8044,7 @@ capacity_from_percent(char *buf)
 					     &req.percent);
 		if (req.ret)
 			return req;
-		if ((u64)req.percent > UCLAMP_PERCENT_SCALE) {
+		if (req.percent > UCLAMP_PERCENT_SCALE) {
 			req.ret = -ERANGE;
 			return req;
 		}
